Handle 64-bit double throughout most of the library
This patch was submitted by Georg-Johann Lay.

It is basically the patch from savannah bug #57071 / issue #670.

Johann's Changelog is:

        Fix for bug #57071: Provide proper double prototypes without __asm.
        Provide long double prototypes.  Provide alias symbols for double
        math functions if double = float.  Provide alias symbols for long
        double math function if long double = float.  Fix places that used
        double to use float instead.  Handle float formats in printf
        depending on whether double = float and / or long double = float.
/
        * include/math.h (__ASM_ALIAS): Remove macro definition.
        (cos, sin, tan, fmod, modf, sqrt, cbrt, hypot, square)
        (floor, ceil, frexp, ldexp, exp, cosh, sinh, tanh)
        (acos, asin, atan, atan2, log, log10, pow, isnan, isinf)
        (signbit, fdim, fma, fmax, fmin, trunc, round, lround)
        (lrint): Provide as vanilla prototypes without __ASM_ALIAS.
        (cosl, sinl, tanl, fmodl, modfl, sqrtl, cbrtl, hypotl, squarel)
        (floorl, ceill, frexpl, ldexpl, expl, coshl, sinhl, tanhl)
        (acosl, asinl, atanl, atan2l, logl, log10l, powl, isnanl, isinfl)
        (signbitl, fdiml, fmal, fmaxl, fminl, truncl, roundl, lroundl)
        (lrintl): New long double prototypes.
        (fabsl, copysignl): New inline functions.
        (isfinite) [double=float]: New inline function.
        (isfinite) [double!=float]: New prototype.
        (isfinitel) [long double=float]: New inline function.
        (isfinitel) [long double!=float]: New prototype.
        (isfnintef): Early-clobber asm output.

        * include/time.h (solar_declinationf, solar_declinationl): New
        prototypes.

        * include/stdlib.h (__ATTR_ALWAYS_INLINE__): New macro.
        (abs, labs): Turn from macros to (inline) functions.
        (strtof, strtold, atoff, atofl, ftostre, ldtostre)
        (ftostrf, ldtostrf): New prototypes.

        * common/alias.h (DALIAS, LALIAS): New macros.

        * common/ftoa_engine.h (__ftoa_engine): Use float instead of double.

libm/fplib/
        * asmdef.h (ALIAS_ENTRY): Remove .macro.
        (ENTRY_FLOAT): New .macro.

        * floor.S (floor, floorl): Provide as alias for the float case
        if double=float resp. long double=float.
        * modf.S (modf, modfl): Same.
        * round.S (round, roundl): Same.
        * fdim.S (fdim, fdiml): Same.
        * fmod.S (fmod, fmodl): Same.
        * fmin.S (fmin, fminl): Same.
        * fmax.S (fmax, fmaxl): Same.
        * sqrt.S (sqrt, sqrtl): Same.
        * cbrt.S (cbrt, cbrtl): Same.
        * log.S (log, logl): Same.
        * log10.S (log10, log10l): Same.
        * exp.S (exp, expl): Same.
        * pow.S (pow, powl): Same.
        * sinh.S (sinh, sinhl): Same.
        * cosh.S (cosh, coshl): Same.
        * tanh.S (tanh, tanhl): Same.
        * sin.S (sin, sinl): Same.
        * cos.S (cos, cosl): Same.
        * tan.S (tan, tanl): Same.
        * atan2.S (atan2, atan2l): Same.
        * atan.S (atan, atanl): Same.
        * asin.S (asin, asinl): Same.
        * acos.S (acos, acosl): Same.
        * fma.S (fma, fmal): Same.
        * ceil.S (ceil, ceill): Same.
        * trunc.S (trunc, truncl): Same.
        * hypot.S (hypot, hypotl): Same.
        * isinf.S (isinf, isinfl): Same.
        * isnan.S (isnan, isnanl): Same.
        * isfinite.S (isfinite, isfinitel): Same.
        * copysign.S (copysign, copysignl): Same.
        * signbit.S (signbit, signbitl): Same.
        * lround.S (lround, lroundl): Same.
        * square.S (square, squarel): Same.
        * frexp.S (frexp, frexpl): Same.
        * ldexp.S (ldexp, ldexpl): Same.
        * lrint.S (lrint, lrintl): Same.

libc/stdlib/
        * dtoa_conv.h (ftoa_prf): Renamed from dtoa_prf, but use float.
        (dtoa_lim, dtoa_cln, d2stream): Remove unused protos.

        * dtoa_prf.c (alias.h): Include it.
        (ftoa_prf): Renamed from dtoa_prf, but use float.
        (dtoa_pr) [double=float]: Alias for ftoa_prf.
        (ldtoa_prf) [long double=float]: Alias for ftoa_prf.

        * atof.S (alias.h): Include it.
        (atoff): Renamed from atof.
        (atof) [double=float]: Alias for atoff.
        (atofl) [long double=float]: Alias for atoff.

        * strtod.c: Use float instead of double.
        (alias.h): Include it.
        (INFINITYf, NANf): New macros using GCC built-ins.
        (strtof): Renamed from strtod.
        (strtod) [double=float]: Alias for strtof.
        (strtold) [long double=float]: Alias for strtof.

        * dtostre.c: Use float instead of double.
        (alias.h): Include it.
        (ftostre): Renamed from dtostre.
        (dtostre) [double=float]: Alias for ftostre.
        (ldtostre) [long double=float]: Alias for ftostre.
        * dtostrf.c (ftostrf, dtostrf, ldtostrf): Similar.

libc/stdio/
        * vfprintf.c (FL_LONG_DOUBLE): New define.
        (vfprintf): If the argument (by means of %-char or by means of
        argument promotion) has more than 4 bytes, then skip appropriate
        number of bytes and output "?".

libc/time/
        * daylight_seconds.c: Use float instead of double.
        * solar_declination.c (solar_declinationf): Renamed from
        solar_declination.
        (solar_declination) [double=float]: Alias for solar_declinationf.
        (solar_declinationl) [long double=float]: Alias for solar_declinationf.

        * libc/stdlib/abs.c: Implement abs without using identifier abs.
        * libc/stdlib/absl.c: Similar for labs.
        * libc/stdlib/Files.am (AVRLIB_CFLAGS) [malloc.o]:
        Add -Wno-maybe-uninitialized to fix warnings.
        [qsort.o]: Add -Wno-sign-compare to fix warnings.
        * include/ctype.h (__ATTR_CONST__): Remove macro.
        (isalnum, isalpha, isascii, isblank, iscntrl, isdigit, isgraph)
        (islower, isprint, ispunct, isspace, isupper, isxdigit)
        (toascii, tolower, toupper): Remove attribute "const", which
        conflicts with GCC's attribute "pure" for these built-ins.

+++ common/ftoa_engine.h
@@ -33,7 +33,7 @@
 
 #ifndef	__ASSEMBLER__
 
-int __ftoa_engine (double val, char *buf,
+int __ftoa_engine (float val, char *buf,
                    unsigned char prec, unsigned char maxdgs);
 
 #endif
